tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/script-language-type.html

<!DOCTYPE html>
<title>Script: combinations of @type and @language</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#prepare-a-script">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var run = false;
</script>
<script type="" language="foo">
run = true;
</script>
<script>
test(function() {
  assert_equals(run, true);
}, "A script with empty type should run");
</script>
