LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/script-language-type-expected.txt


FAIL A script with empty type should run assert_equals: expected true but got false
